+++ src/domain/aggregate/member.rs
@@ -6,14 +6,14 @@ use super::value_object::{grade::Grade, major::Major};
 pub struct Member {
     pub id: MemberId, // メンバーのID (Value Object)
     pub name: String,
-    pub age: usize,
+    pub age: i32,
     pub grade: Grade,
     pub major: Major,
 }
 
 impl Member {
     // メンバーの新規作成メソッド
-    pub fn new(name: String, age: usize, grade: Grade, major: Major) -> Self {
+    pub fn new(name: String, age: i32, grade: Grade, major: Major) -> Self {
         Member {
             id: MemberId::gen(),
             name,
@@ -24,7 +24,7 @@ impl Member {
     }
 
     // メンバーの再構成メソッド
-    pub fn reconstruct(id: MemberId, name: String, age: usize, grade: Grade, major: Major) -> Self {
+    pub fn reconstruct(id: MemberId, name: String, age: i32, grade: Grade, major: Major) -> Self {
         Member {
             id,
             name,
